Big API refactor to 0.2.0
BacktrackingIterator is now a trait with associated types.
history_position() -> get_ref_position()
backtrack() -> start_again() (the new definition of backtrack is a
different method)
Walkback no longer has to yield refs, and in the copying case, it
doesn't, because that was slightly nonsensical - we weren't avoiding
copying for a real reason.

The previous BacktrackingIterator -> CopyingBacktrackingIterator, to
make the copying more explicit (and to allow for a zero-copying version
in the future) and implements the new traits.
Introduced a trait Walkbackable, to allow for generifying over the
lifetime of the reference generating the backwalk, as
CopyingBacktrackingIterator has done, using `for<>`.
This also now allows for backtracking iterators which cannot implement a
backwalk.