Added a reference-taking backtracking iterator
BacktrackingRecorder is the base wrapper around an existing iterator to
allow it to be backtracked on, but it has mutability problems since it
allows history manipulation, which would invalidate historical
references generated by some iterator.
ReferencingBacktrackingIterator can be generated from a
BacktrackingRecorder with the `record()` method - it takes a mutable
borrow on the recorder to prevent history manipulation, and uses a small
amount of unsafe code to produce historical references.
Walkbacks are done directly on the BacktrackingRecorder - of course,
they also freeze history for their duration.