Moved forgetting onto the BacktrackingRecord type
The possibility of having `forget` on the backtracking iterator was
causing a potential conflict with the referencing form, which tries to
guarantee that the reference lives as long as the iterator - which a
history edit would clearly invalidate.
Instead, `forget_before` and `forget` have been moved onto a new trait
`Record`, which will begin to absorb more generic behaviours for
backtracking as well. This will be a breaking change to the API.